Move Module & ModuleType models to a separate file
