+++ netbox/dcim/models/__init__.py
@@ -2,6 +2,7 @@ from .cables import *
 from .device_component_templates import *
 from .device_components import *
 from .devices import *
+from .modules import *
 from .power import *
 from .racks import *
 from .sites import *
